chore: move `SanityChecker` into `physical-optimizer` crate (#14083)
* chore: move  into  crate

* chore: move SanityChecker tests out to datafusion/core/tests

* chore: update datafusion-cli/Cargo.lock

* fix cargo doc

---------
