+++ datafusion/core/src/physical_optimizer/mod.rs
@@ -27,7 +27,6 @@ pub mod enforce_sorting;
 pub mod optimizer;
 pub mod projection_pushdown;
 pub mod replace_with_order_preserving_variants;
-pub mod sanity_checker;
 #[cfg(test)]
 pub mod test_utils;
