+++ datafusion/physical-optimizer/src/lib.rs
@@ -27,6 +27,8 @@ pub mod limited_distinct_aggregation;
 mod optimizer;
 pub mod output_requirements;
 pub mod pruning;
+pub mod sanity_checker;
+pub mod test_utils;
 pub mod topk_aggregation;
 pub mod update_aggr_exprs;
